latest

+++ config/mail.php
@@ -38,7 +38,7 @@ return [
             'transport' => 'smtp',
             'url' => env('MAIL_URL'),
             'host' => env('MAIL_HOST'),
-            'port' => env('MAIL_PORT'),
+            'port' => (int) env('MAIL_PORT'),
             'encryption' => env('MAIL_ENCRYPTION'),
             'username' => env('MAIL_USERNAME'),
             'password' => env('MAIL_PASSWORD'),
